==> Notebook/1.1-dataanalysis.ipynb <==
# INFO-F-422 -  Statistical Foundations of Machine Learning 

### Couchard Darious - __[Darius.Couchard@ulb.be](mailto:Darius.Couchard@ulb.be) - Student ID 425366__
### Donne Stefano - __[Stefano.Donne@ulb.be](mailto:Stefano.Donne@ulb.be) - Student ID 408801__
### Parent Paul - __[Paul.Parent@ulb.be](mailto:Paul.Parent@ulb.be) - Student ID 495257__


# 1) Data analysis

All the features will be analysed one by one. Some of them are not representative of the problem and can be ommited.

training_set<-read.csv("../Data/TrainingSet/4910797b-ee55-40a7-8668-10efd5c1b960.csv",header=TRUE) # loads the training set csv file (it's magic)
training_labels<-read.csv("../Data/TrainingLabel/0bf8bc6e-30d0-4c50-956a-603fc693d966.csv", header=TRUE) # Loads the corresponding labels
training_set <- merge(training_set, training_labels)
dim(training_set)
names(training_set)

## Analysis of 18 first features

### Functions used to analyse these features

# Count the number of missing values 
sum(is.na(training_set$amount_tsh))

# Count the number of a specific value (0 for example)
tabl <- table(training_set$amount_tsh)
tabl[names(tabl)=="0"]

# Count number of levels of a vector
length(unique(training_set$amount_tsh))

**id**: Unique value for each pump.

**amount_tsh**: amount of water available to a waterpoint. No information whether 0 is missing value or because there is no water available.

**date_recorded**: the date the row was recorded in the file. Irrelevant for the functionnality.

**funder**: (4412 NA (3635 "blanks" + 777 "0")) who funded wells. 1897 funders in total. Divide them in 2 categories of funder?

**gps_height**: "0" values are probably NA (21960). Too much NA to take this feature into account?

**installer**: (4432 NA (3655 "blanks" + 777 "0")) who installed wells. 2146 installers. Same as funder?

**longitude-latitude**: (No NA) Geographical feature.

**wpt_name**: Name of the waterpoint. Irrelevant for the functionnality.

**num_private**: 95% of NA. Could not find signification of this feature.

**basin**: (No NA) Geographical feature. 9 lakes.

**subvillage**: (No NA) Geographical feature. 19287 different villages. Too much unique values? 

**region**: (No NA) Geographical feature. 21 regions.

**region_code**: (No NA). Geographical feature. 28 region codes.

**lga**: (No NA) Geographical feature. 125 different lga.

**ward**: (No NA) Geographical feature. 2091 different wards.

**population**: 22900 value of "0" (consider "0" as NA or no population?). Take mean of the region to impute NA?

**public_meeting**: Irrelevant for the functionnality.

**recorded_by**: Irrelevant for the functionnality.

## Analysis of 20 last features

### Functions used to analyze these features

# List of categories for a column and count their occurences

count_categories <- function (data, column) {
    temp = data.frame(data["id"], data[column], data["status_group"])
    temp_names = as.data.frame(table(temp[column]))
    names(temp_names)[1] <- column
    return(temp_names)
}

count_categories(training_set, "scheme_management")

# Count of status_group (functional, non functional or functional needs repairs) for each category of a column

library(dplyr)

count_status <- function(data, column) {
    number_categories <- dim(unique(data[column]))[1]
    res_matrix <- matrix(,4, number_categories)  # Create matrix with 3 rows (labels) and N cols (number of categories)
        
    count <- 1
    for (category in unique(data[,column])) {
        filtered <- filter(data, data[,column] == category)
        temp_categories <- count_categories(filtered, "status_group")
        for (i in 1:3) {
            res_matrix[i, count] <- temp_categories$Freq[i]
        }
        res_matrix[4, count] <- res_matrix[1, count] / (res_matrix[1, count] + res_matrix[2, count] + res_matrix[3, count])
        count <- count + 1
    }
    res_matrix[is.na(res_matrix)] <- 0
    rownames(res_matrix) <- c("functional", "need repair", "non functional", "func %")
    colnames(res_matrix) <- unique(data[,column])
    return (res_matrix)
}

### scheme_management

# only print 2 digits
options(digits=2)

count_status(training_set, "scheme_management")

Scheme management is rather uncorrellated with the status group.

### management

count_status(training_set, "management")

Management is also an important class, maybe management group is better?

### scheme_name

count_status(training_set, "scheme_name")

Scheme name has too much NA values and different categories. This cannot be used by our model

### permit 

count_status(training_set, "permit")

Permit is lightly correllated with the status group. This measure need to be kept and encoded into an ont-hot encoded matrix. How do we treat N/A values? => Use the mean of true/false after standardization

### construction_year 

count_status(training_set, "construction_year")

Construction year is an important measure, need to be kept and encoded as an integer (by the pump age). Many NA values exists and should be dealt with.

### extraction_type

count_status(training_set, "extraction_type")

 Extraction type also looks like an importan measure, as it influences the operating mode of the pump. Some classes have better results than others. This should also be encoded into a one-hot matrix (but has many dimensions)

### extraction_type_group

count_status(training_set, "extraction_type_group")

Perhaps it is better to use the group than the extraction type, as it has less categories but still shows differences between the categories.

### extraction_type_class

count_status(training_set, "extraction_type_class")

Using a type class also has again less categories than the group and shows differences between the categorieds.

Ultimately only one of those three categories should be used.

### management_group

count_status(training_set, "management_group")

Management group is less corellated than management

### payment

count_status(training_set, "payment")

Payment looks like an important value, (can maybe be encoded into an integer?)

### payment_type

count_status(training_set, "payment_type")

Duplicate column from payment, can be discarded

### water_quality

count_status(training_set, "water_quality")

Might be an interesting data but must be processed (maybe encoded into an integer again?)

### quality_group

count_status(training_set, "quality_group")

Better than water_quality, should be used an encoded as integer.

### quantity

count_status(training_set, "quantity")

Important, should be kept.

### quantity_group

count_status(training_set, "quantity_group")

duplicate, can be dropped

### source

count_status(training_set, "source")

Important, should be kept

### source_type

count_status(training_set, "source_type")

Perhaps better than previous.

### source_class

count_status(training_set, "source_class")

Not much correlation, should be dropped.

### waterpoint_type

count_status(training_set, "waterpoint_type")

High correlation, should be kept

### waterpoint_type_group

count_status(training_set, "waterpoint_type_group")

Again a more compact still interesting feature

## To summerize the different features
* **Scheme management** is uncorrelated.
* **Scheme name** have too much unique names and NA values, not interesting.
* **Permit** is important (impact on reparation) and should be one-hot encoded.
* **Construction** year looks important and should be changed to age.
* **Extraction type**, extraction type group and extraction type class are important, one of them three should be kept.
* **Management** and **management group** are important, one of both should be kept.
* **payment* is important, maybe can be encoded as integer.
* **payment type** is DUPLICATE.
* **water quality** and **quality group** are also important, one of both should be kept, maybe can be encoded as integer.
* **water quantity** is important, one should be kept, maybe encoded as integer.
* **water quantity group** is DUPLICATE.
* **source** and **source type**, one should be kept. One hot encoded.
* **source class* is rather uncorellated. Should be dropped.
* **waterpoint type** and **waterpoint type group** are importnat, one should be kept.

* **Scheme management** => One hot encoding, fill N/A
* **Permit** => Fill N/A
* **Construction year** => change to age, fille N/A with mean after standardisation
* **Extraction type class** => One hot
* **Management group** => One hot
* **Payment** => One hot
* **Water quality group** => CAN be put as an integer
* **Water quantity** => CAN be put as an integer
* **Source type** => One hot
* **Waterpoint type group** => One hot
