==> Notebook/1.3-datasetsplit.ipynb <==
# INFO-F-422 -  Statistical Foundations of Machine Learning 

### Couchard Darious - __[Darius.Couchard@ulb.be](mailto:Darius.Couchard@ulb.be) - Student ID 425366__
### Donne Stefano - __[Stefano.Donne@ulb.be](mailto:Stefano.Donne@ulb.be) - Student ID 408801__
### Parent Paul - __[Paul.Parent@ulb.be](mailto:Paul.Parent@ulb.be) - Student ID 495257__


# 3) Dataset split

The dataset will be split into a training set (around 75-80%) and a validation set (around 25-20%).

Creating a validation dataset is a measuring tool against overfitting. Obtaining a good accuracy on the training set but a bad accuracy on the validation set is sign of overfitting.

In order to correctly compare different models (Random Forest, SVM, etc...) or slightly different dataset (e.g. we drop some rows), we need our split to be deterministic for each row.

# https://github.com/ropensci/writexl
library(nycflights13)
# Read the preprocessed training data
dataset <- read.csv("../Data/PreProcess/processed_training_data.csv")

dim(dataset)

# Random seed
set.seed(42) # The Answer to the Ultimate Question of Life, the Universe, and Everything is 42

ind <- sample(2, nrow(dataset), replace=TRUE, prob=c(0.8, 0.2))
training_dataset <- dataset[ind==1,]
verification_dataset <- dataset[ind==2,]

dim(training_dataset)[1] + dim(verification_dataset)[1]

write.csv(training_dataset, "../Data/PreProcess/processed_training_data_split.csv")
write.csv(verification_dataset, "../Data/PreProcess/processed_verification_data_split.csv")
