==> Notebook/2.2-SVM.ipynb <==
# Model Selection : Support Vector Machine (SVM) Classifier

Consists of creating an hyperplane that divides the instances in space.

This method is effective with highly dimentional problems.

library(e1071)
training_set <- read.csv("../Data/PreProcess/processed_training_data_split.csv")
validation_set <- read.csv("../Data/PreProcess/processed_verification_data_split.csv")

column_to_drop<-c("X.1","X")
training_set<-training_set[,!(names(training_set) %in% column_to_drop)] # drop the desired columns
validation_set<-validation_set[,!(names(validation_set) %in% column_to_drop)] # drop the desired columns

training_set$id <- factor(training_set$id)
training_set

## SVM Hyperparameters

The svm classifier has multiples parameters such as:

* Kernel type, linear, polynomial, radial_based, sigmoid. Changes the shape of the diving surface
* Degree, needed if polynomial.
* Gamma, affects the shape of the surface, needed if not linear
* coef0, needed in polynomial and sigmoid
* cost, cost of constraint violation

The SVM classifier is trainined on the training set, accuracy is then measured with the validation set.

svm_model <- svm(id ~ ., data=training_set, kernel="linear", cost=10, scale=FALSE)  # No need to scale as data is already processed
print(svm_model)

## Training set accuracy

svm.predict <- predict(svm_model, training_set[1:100,])

classAgreement(table(pred=svm.predict, true=training_set$id[1:100]))

## Validation set accuracy

svm.predict <- predict(svm_model, validation_set)

classAgreement(table(pred=svm.predict, true=validation_set$id[-1]))

## Testing with polynomial SVM

svm_model_poly <- svm(id ~ ., data=training_set, kernel="polynomial", cost=20, scale=FALSE)

svm.predict <- predict(svm_model_poly, training_set[1:500,])
classAgreement(table(pred=svm.predict, true=training_set$id[1:500]))

svm.predict <- predict(svm_model_poly, validation_set)
classAgreement(table(pred=svm.predict, true=validation_set$id[-1]))

Increasing cost increases the accuracy of both training and validation set, but also increases overfitting.
* With cost=20, training_accuracy = 72% & validation_accuracy=66%

## Testing with radial basis SVM 

svm_model_poly <- svm(id ~ ., data=training_set, kernel="radial", cost=10, scale=FALSE)

svm.predict <- predict(svm_model_poly, training_set[1:500,])
classAgreement(table(pred=svm.predict, true=training_set$id[1:500]))

svm.predict <- predict(svm_model_poly, validation_set)
classAgreement(table(pred=svm.predict, true=validation_set$id[-1]))

## Fixing multiclass classification with SVM

svm_model_poly <- svm(id ~ ., data=training_set, kernel="polynomial", cost=10, scale=FALSE, probability=TRUE)

svm.predict <- predict(svm_model_poly, validation_set, probability=TRUE, decision.values=TRUE)

classAgreement(table(pred=svm.predict, true=validation_set$id[-1]))

library(caret)
confusionMatrix(data=svm.predict, reference=factor(validation_set$id[-1]))

Multiclass classification is now working, but only few instances are classifier as "functional need repairs"

## Performing Grid search for the best parameters of the SVM with polynormial kernel

cost_list <- c(1, 5, 10)  # Too much overfitting after 10
#coef0_list <- c(0.0, 1.0) # turn that off to start with
degree_list <- c(3, 5, 7)  # Coefficient degrees -> too much overfitting can happen
accuracy_matrix <- matrix(0, ncol=3, nrow=3, dimnames=list(degree_list, cost_list))

for (i in 1:length(cost_list)) {
    for (j in 1:length(degree_list)) {
        cost_ <- cost_list[i]
        deg_ <- degree_list[j]
        svm_model_poly <- svm(id ~ ., data=training_set, kernel="polynomial", cost=cost_, degree=deg_, scale=FALSE)
        svm.predict <- predict(svm_model_poly, validation_set)
        ag <- classAgreement(table(pred=svm.predict, true=validation_set$id[-1]))
        accuracy_matrix[j, i] <- ag$diag
        print(ag$diag)
    }
}

accuracy_matrix

save(svm_model_poly, file="4-Models/svm_poly_classifier.RData")
