==> Notebook/2.3-nnet.ipynb <==
# Model Selection : Neural network (nnet)

library(nnet)
library(caret)

# Confusion matrix function
confusion_matrix <- function(valid, predi, posLabel="yes"){
 cm <- table(valid, predi)
 print("Confusion matrix")
 print(cm)
 error_rate <- 1-sum(diag(cm))/sum(cm)
 print(paste("Error rate =", error_rate))
}

# Data
training_set <- read.csv("../Data/PreProcess/processed_training_data_split.csv")
validation_set <- read.csv("../Data/PreProcess/processed_verification_data_split.csv")

training_set

# Columns to drop in training and validation dataset
columns_to_drop<-c("X.1","X")
training_set<-training_set[,!(names(training_set) %in% columns_to_drop)] # Drop the desired columns
validation_set<-validation_set[,!(names(validation_set) %in% columns_to_drop)]# Drop the desired columns

training_set$id = factor(training_set$id)

### Neural network with hidden layer (11 neurons)

# Neural network with 11 neurons in hidden layer
nn11 <- nnet(id ~ ., data = training_set, skip = FALSE, size = 11, maxit = 1000)

# Confusion matrix
pred_nn11 <- predict(nn11, newdata = validation_set, type = "class")

confusionMatrix(table(pred_nn11, validation_set$id))

### NN runs:
- No hidden layer: 0.31 error rate
- 2 neurons hidden layer: 0.49 error rate
- 11 neurons hidden layer: 0.25 error rate

save(nn11, file="4-Models/nn_model_classifier.RData")
