==> Notebook/3.1-GradientBoosting.ipynb <==
# Gradient Boosting classifier with decision trees

Gradient boosting is an ensemble learning technique that consists of sequentially adding predictors. Each predictor will fit on the residual error of the previous predictor.

The predictors used in this case are weak learners decision trees.

library(xgboost)
library(caret)

column_to_drop<-c("X.1","X")
training_set <- read.csv("../Data/PreProcess/processed_training_data_split.csv") # load training set
validation_set <- read.csv("../Data/PreProcess/processed_verification_data_split.csv") # load validation set

# Dropped "X.1","X" because they just represent the row numbers
training_set<-training_set[,!(names(validation_set) %in% column_to_drop)] # drop the desired columns on validation set
validation_set<-validation_set[,!(names(validation_set) %in% column_to_drop)] # drop the desired columns on validation set

## test GBM - fait planter le kernel 
library(gbm)
gbm.fit <- gbm(
  formula = id ~ .,
  distribution = "gaussian",
  data = training_set,
  n.trees = 10,
  interaction.depth = 1,
  shrinkage = 0.9,
  cv.folds = 5,
  verbose = FALSE
  )  

## test xgboost
modelGB <- train(
  id ~., data = training_set, method = "xgbTree",
  trControl = trainControl("cv", number = 10)
  )
# tuning parameter
modelGB$bestTune

modelGB
save(modelGB,file = "4-Models/modelGB.RData")

load("4-Models/modelGB.RData")
pred <- predict(modelGB,validation_set)
conf <- confusionMatrix(data=pred, reference = factor(validation_set$id[-1])) # GBM confusion matrix
conf
