==> Notebook/4-CompetitionPrediction.ipynb <==
# INFO-F-422 -  Statistical Foundations of Machine Learning 

### Couchard Darious - __[Darius.Couchard@ulb.be](mailto:Darius.Couchard@ulb.be) - Student ID 425366__
### Donne Stefano - __[Stefano.Donne@ulb.be](mailto:Stefano.Donne@ulb.be) - Student ID 408801__
### Parent Paul - __[Paul.Parent@ulb.be](mailto:Paul.Parent@ulb.be) - Student ID 495257__


## 4) In this section the best model obtained so far is used to predict labels on the test set of the DrivenData competition

library(randomForest)
library(caret)
test_set <- read.csv("../Data/PreProcess/test_data.csv")

load("4-Models/random_forest.RData") # load Random Forest trained model -> name : classifierRF
pred <- predict(classifierRF,test_set, type = "class") # Random Forest model prediction on test sett

output <- data.frame(id=test_set$id,status_group=pred)
output

write.csv(output,"../Data/TestOutput/output.csv", row.names=FALSE)
