==> Notebook/4.3-nnet.ipynb <==
## Neural network (nnet)

We used the lirabry nnet to train a Neural Network (NN) model.

library(nnet)
library(caret)

# Data
training_set <- read.csv("../Data/PreProcess/processed_training_data_split.csv")
validation_set <- read.csv("../Data/PreProcess/processed_verification_data_split.csv")

training_set

# Columns to drop in training and validation dataset
columns_to_drop<-c("X.1","X")
training_set<-training_set[,!(names(training_set) %in% columns_to_drop)] # Drop the desired columns
validation_set<-validation_set[,!(names(validation_set) %in% columns_to_drop)]# Drop the desired columns

training_set$id = factor(training_set$id)

### Neural network with hidden layer

# Neural network with 5 neurons in hidden layer
nn5 <- nnet(id ~ ., data = training_set, skip = FALSE, size = 5, maxit = 1000)

# Confusion matrix
pred_nn5 <- predict(nn5, newdata = validation_set, type = "class")

confusionMatrix(factor(pred_nn5), factor(validation_set$id))

# Neural network with 8 neurons in hidden layer
nn8 <- nnet(id ~ ., data = training_set, skip = FALSE, size = 8, maxit = 1000)

# Confusion matrix
pred_nn8 <- predict(nn8, newdata = validation_set, type = "class")

confusionMatrix(factor(pred_nn8), factor(validation_set$id))

# Neural network with 11 neurons in hidden layer
nn11 <- nnet(id ~ ., data = training_set, skip = FALSE, size = 11, maxit = 10000)

# Confusion matrix
pred_nn11 <- predict(nn11, newdata = validation_set, type = "class")

confusionMatrix(factor(pred_nn11), factor(validation_set$id))

# Neural network with 15 neurons in hidden layer
nn15 <- nnet(id ~ ., data = training_set, skip = FALSE, size = 15, maxit = 10000, MaxNWts = 10000) #Increase of iterations and maximum weights

# Confusion matrix
pred_nn15 <- predict(nn15, newdata = validation_set, type = "class")

confusionMatrix(table(pred_nn15, validation_set$id))

# Neural network with 20 neurons in hidden layer
nn20 <- nnet(id ~ ., data = training_set, skip = FALSE, size = 20, maxit = 10000, MaxNWts = 10000) #Increase of iterations and maximum weights

# Confusion matrix
pred_nn20 <- predict(nn20, newdata = validation_set, type = "class")

confusionMatrix(table(pred_nn20, validation_set$id))

save(nn20, file="4-Models/nn_model_classifier.RData")
